Issue #32: View Earned Achievements: Code Reorganisation + Earned achievements link
Reorganised code.

Added a link to the 'earned achievements' page.
